Split `value-visit.ts` into a submodule with multiple files.
