+++ packages/data-model/deno.jsonc
@@ -6,6 +6,7 @@
   "imports": {
     // Dependency guide: ../../docs/development/DEPENDENCIES.md
     "@/": "./src/",
+    "@/value-visit": "./src/value-visit/index.ts",
     "@node/crypto": "node:crypto"
   },
   "tasks": {
